Skin import working again :white_check_mark:
The skin import is now working again after I've messed around with sockets and removed it.

In the next update I'm going to make it independent of socket.io and remove the module..

The problem with socket.io is that it can cause problems if the ports that the app is uing is blocked. It also has some security issues.